+++ packages/giselle/src/engine/acts/index.ts
@@ -2,6 +2,5 @@ export * from "./create-act";
 export * from "./create-and-start-act";
 export * from "./get-act";
 export * from "./get-workspace-acts";
-export * from "./object";
 export * from "./patch-act";
 export * from "./start-act";
