--- packages/giselle/src/engine/acts/object/index.ts
@@ -1,2 +1,0 @@
-// Re-export all Act-related types from concepts
-export * from "../../../concepts/act";
