+++ packages/giselle/src/engine/acts/patch-act.ts
@@ -1,6 +1,6 @@
 import type { ActId } from "../../concepts/identifiers";
 import type { GiselleEngineContext } from "../types";
-import type { Act } from "./object";
+import type { Act } from "../../concepts/act";
 import {
 	type PatchDelta,
 	patchAct as patchActObject,
