+++ packages/giselle/src/engine/acts/start-act.ts
@@ -13,7 +13,7 @@ import {
 import { executeAction } from "../operations";
 import { executeQuery } from "../operations/execute-query";
 import type { GiselleEngineContext } from "../types";
-import { Act, type Sequence } from "./object";
+import { Act, type Sequence } from "../../concepts/act";
 import { actPath } from "./object/paths";
 import { patchAct } from "./patch-act";
